+++ modules/runner-binaries-syncer/variables.tf
@@ -49,10 +49,10 @@ variable "s3_logging_bucket_prefix" {
   type        = string
   default     = null
 
-  # Make sure the bucket name only contains legal characters
+  # Make sure the bucket prefix only contains legal characters
   validation {
-    error_message = "Only lowercase alphanumeric characters and hyphens allowed in the bucket name."
-    condition     = var.s3_logging_bucket_prefix == null || can(regex("^[a-z0-9-]*$", var.s3_logging_bucket_prefix))
+    error_message = "Only alphanumeric characters, hyphens followed by single slashes allowed in the bucket prefix."
+    condition     = var.s3_logging_bucket_prefix == null || can(regex("^(([a-zA-Z0-9-])+(\\/?))*$", var.s3_logging_bucket_prefix))
   }
 }
 
@@ -206,7 +206,7 @@ variable "log_level" {
 }
 
 variable "server_side_encryption_configuration" {
-  description = "Map containing server-side encryption configuration."
+  description = "Map containing server-side encryption configuration for runner-binaries S3 bucket."
   type        = any
   default = {
     rule = {
@@ -215,6 +215,12 @@ variable "server_side_encryption_configuration" {
       }
     }
   }
+}
+
+variable "s3_versioning" {
+  description = "Status of S3 versioning for runner-binaries S3 bucket."
+  type        = string
+  default     = "Disabled"
 }
 
 variable "lambda_principals" {
@@ -240,4 +246,10 @@ variable "lambda_architecture" {
     condition     = contains(["arm64", "x86_64"], var.lambda_architecture)
     error_message = "`lambda_architecture` value is not valid, valid values are: `arm64` and `x86_64`."
   }
+}
+
+variable "lambda_tracing_mode" {
+  description = "Enable X-Ray tracing for the lambda functions."
+  type        = string
+  default     = null
 }